added the APIs using swagger
