+++ Dockerfile
@@ -6,6 +6,7 @@ WORKDIR /usr/src/relationships_g3
 
 # Install app dependencies
 COPY package.json /usr/src/relationships_g3/
+RUN npm install -g swagger
 RUN npm install
 # Bundle app source
 COPY . /usr/src/relationships_g3
