+++ algorithms/ollie/docker-compose.yml
@@ -1,0 +1,7 @@
+version: "2"
+services:
+  ollie:
+    build: .
+    ports:
+    - "3000:3000"
+
