+++ algorithms/ollie/example/input.txt
@@ -1,0 +1,2 @@
+Educated at the Royal College of Music, he began his career contributing songs to revues and incidental music for the stage, forming a notable partnership with the playwright Clemence Dane.
+asdasdsdasda asdascasda adas
