+++ algorithms/openie/api/swagger/swagger.yaml
@@ -1,0 +1,93 @@
+---
+swagger: "2.0"
+info:
+  description: "Get the Music relationships by running the OpenIe Algorithm on a set\
+    \ of sentences."
+  version: "1.0.0"
+  title: "OpenIE Algorithm - RelationshipG3 API"
+
+basePath: "/relationshipg3/Openie"
+schemes:
+- "http"
+securityDefinitions:
+  basicAuth:
+    type: basic
+    description: HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
+produces:
+- "application/json"
+paths:
+  /getRelationships:
+    get:
+      security:
+      - basicAuth: []
+      tags:
+      - "GetRelationships"
+      summary: "Get the relationships from the given set of text array"
+      description: "The GetRelationships endpoint returns all the relationships found\
+        \ in the text by running openIE algorithm on it. The response includes the\
+        \ sentence and the instances which contains detail about each instance of\
+        \ relationship found. \n"
+      operationId: "getRelationshipsGET"
+      parameters:
+      - name: "inputTextArray"
+        in: "query"
+        description: "The input text array containing the sentences"
+        required: false
+        type: "array"
+        items: {
+          type: "string"
+        }
+      - name: "inputFilePath"
+        in: "query"
+        description: "The input filepath consisting of set of sentences"
+        required: false
+        type: "string"
+        format: "url"
+      responses:
+        200:
+          description: "An array of Relationships"
+          schema:
+            type: "array"
+            items:
+              $ref: "#/definitions/Relationships"
+        default:
+          description: "Unexpected error"
+          schema:
+            $ref: "#/definitions/Error"
+      x-swagger-router-controller: "GetRelationships"
+definitions:
+  Relationships:
+    type: "object"
+    properties:
+      sentence:
+        type: "string"
+        description: "Unique sentence present in the given text."
+      instances:
+        type: "array"
+        description: "Relation instances present in the sentence."
+        items:
+          $ref: "#/definitions/Instance"
+  Instance:
+    type: "object"
+    properties:
+      quality:
+        type: "number"
+        description: "a quality given to the relationship found"
+      term1:
+        type: "string"
+        description: "Entity or a relation present in the sentence"
+      relation:
+        type: "string"
+      term2:
+        type: "string"
+        description: "Entity present in the sentence"
+  Error:
+    type: "object"
+    properties:
+      code:
+        type: "integer"
+        format: "int32"
+      message:
+        type: "string"
+      fields:
+        type: "string"
