+++ algorithms/openie/example/input.txt
@@ -1,0 +1,3 @@
+hello i am here
+hello i am here
+whats is going on
