+++ algorithms/openie/example/test.txt
@@ -1,0 +1,9 @@
+Educated at the Royal College of Music, he began his career contributing songs to revues and incidental music for the stage, forming a notable partnership with the playwright Clemence Dane.
+He is best remembered now as a composer for British cinema, his career in that medium beginning in 1936 and achieving early and widespread recognition with his score for the Oscar-winning Goodbye, Mr Chips in 1939.
+He went on to work in revue with the legendary comedienne Joyce Grenfell (1910--79), writing songs with her for West End shows like Tuppence Coloured (1947) and Penny Plain (1951).
+Undoubtedly his most successful work was to be the Warsaw Concerto, for piano and orchestra in the grand heroic style of Rachmaninov, the most memorable feature of the film Dangerous Moonlight (1941).
+But his fluent and versatile writing was to prove highly suitable to a whole era of British films of the mid-twentieth century, in many instances, as in all the best film scores, contributing independently to the popular success of the film.
+Addinsell was a match for many cinematic genres: historic drama (Fire over England, 1937, Tom Brown's Schooldays, 1950, Beau Brummel, 1954)), psychological (Gaslight, 1940), contemporary (Love on the Dole, 1941, Life at the Top, 1965) or even comedy (The Prince and the Showgirl, 1957, Waltz of the Toreadors, 1962).
+He was hugely influential on a generation of British film composers and established a quality and style of full scale orchestral writing that was never bettered.
+
+
