--- app/algorithms/ollie/config/index.js
@@ -1,14 +1,0 @@
-// config.js
-module.exports = {
-  "server":{
-    "port": "3000",
-    "host": "localhost"
-  },
-  "ollieAlgo":
-    {
-      "name":"ollie-app-latest.jar",
-      "javaOpt":"-Xmx512m",
-      "defaultFilePath": "example/test.txt",
-      "defaultFileSavePath": "example/input.txt"
-    }
-};
