--- app/algorithms/ollie/example/input.txt
@@ -1,2 +1,0 @@
-hello i am here
-whats is going on
