--- app/algorithms/openie/docker-compose.yml
@@ -1,7 +1,0 @@
-version: "2"
-services:
-  openie:
-    build: .
-    ports:
-    - "3001:3001"
-
