--- app/algorithms/openie/example/input.txt
@@ -1,3 +1,0 @@
-hello i am here
-hello i am here
-whats is going on
