--- app/algorithms/openie/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "openie",
-  "version": "1.0.0",
-  "description": "node.js wrapper for openie algorithm",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.17.1",
-    "child_process": "^1.0.2",
-    "clone": "^2.1.1",
-    "express": "^4.15.2",
-    "fs": "0.0.1-security",
-    "morgan": "^1.8.1",
-    "request": "^2.81.0",
-    "tail": "^1.2.1"
-  },
-  "devDependencies": {}
-}
